Inline Shell startup helpers
